Introduce hyperlight library as an external dependency for Envoy
Hyperlight is using Rust and Cargo. Additionally Hyperlight
depends on some private Rust crate registries that require
authorization to download crates.

It's is in principle possible to depend on Rust/Cargo repositories
in Bazel using rules_rust package. However, when it comes to
handling transitive dependencies of Rust/Cargo repository,
rules_rust does not seem to support authorization and thus cannot
work with private repositories that require authorization.

So even though we can generate all the relevant Bazel rules to
download hyperlight and its dependencies, those rules will not be
able to download anything.

So for now, until the issues are address, I add statically linked
libnrary as an external dependency to Envoy. Moreover, I expect
this statically linked library to be stored in the local filesystem.

NOTE: That's just a personal prefernece, I could have uploaded the
static library directly to github, but I don't like the idea of
storing build artifacts in the VCS - VCS should be for code and
build instructions/scripts.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>