Wrap hyperlight library into a C++ interface
This is really a rather quickly and mindlessly hacked
up wrapper, just to be able to use the hyperlight library
a bit easier.

One particular issue with the current implementation (besides
the fact that there is no error handling and no configuration/
tunning knobs exposed) is that when you run a Hyperlight
sandbox you also have to retain the Builder that was used to
construct that sandbox. That's quite counterintuitive.

The reason why it currently works this way is that we register
host functions that the WASM guest can call in the builder and
the whole machinery needed to find the right callback when the
guest calls a host function is implemented in the Builder.

Droping the Builder will result in releasing memory used by
those callbacks and all kinds of weird memory corruption bugs.
We deifinetely don't want to keep the current interface as-is
in the real implementation.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>