Implement Hyperlight WASM filter
At this point this filter is not very useful by itself.
However, it is fully functional - it runs a WASM module
inside Hyperlight and that module can call host functions
to do things (only write function is currently provided).

The implementation is rather inefficient and there have
been some dubious decisions made to make it work, you
can refer to extensive comments in the code as well as
previous comments that explain those issues.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>