Actulally consume the data from the buffer in the onData call
Currently the Hyperlight echo filter sends the data it receives
during the onData call back, but never actually removes the data
it received from the buffer.

As a result when the data arrives in multiple chunks, because
it's not being removed from the buffer, some of the data is sent
back multiple times.

This change removes the data from the buffer after it was sent
back, thus avoiding this problem.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>