Modify Envoy to be able to implement echo filter in proxy-wasm
Currently proxy-wasm does not have a function that allows to send responses
directly back to the client, so this change adds such a function to Envoy.

Additionally, I modify the configuration of the proxy-wasm filters to allow
proxy-wasm filters to become terminal (IOW, it would allow us to have a
proxy-wasm filter as a last filter of the filter chain).

I don't actually intend this to be pushed upstream at any point, I'm only
introducing this to be able to compare performance of the hyperlight-wasm
echo filter with the proxy-wasm echo filter implementation. Even if it's
not exactly going to be an apples to apples comparision.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>