Asynchonify findFilterChain and drop consts
Currently findFilterChain is synchronous and is not expected to pause.
However as part of working on https://github.com/envoyproxy/envoy/issues/35958
we want to be able to run listner filters on demand when looking for the
right filter chain to match (you can find more context in the linked issue).

The idea is to create an type of input for the filter chain matchers that
will run the appropriate listener filter on demand if necessary (e.g., if we
didn't run it already) and populate the relevant metadata based on the results.

In order to make this work we need two things:

1. Inputs now can change stream info and socket metadata, so we need to drop
   consts around those and prepare that they might change during the course
   of filter chaing matching
2. We need to change the findFilterChain so that it can pause and resume, because
   listener filters may pause.

So this change tries to do all of that without breaking the existsing tests.
Removing the consts seems more or less trivial, so I think the code is self
explanatory.

Supporting pausing/resuming in findFilterChain basically boils down to changing
the API from returning a filter chain directly to calling a callback once the
filter chain has been found.

We will need to make some additional bits and pieces available to the inputs
themselves, so that they can pause (e.g., they might need access to the working
thread dispatcher), but that will be in a sperate change.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>