Fix hessian2_codec Envoy dependency builds in Dalec
It's not immediately clear why in Dalec hessian2_codec does not build
successfully, while using Envoy CI Docker image it builds just fine.

However, it's pretty clear to me that hessian2 repo contains an issue
with Bazel build rules and fixing it resolves the Dalec build problem.

I've sent a patch to the upstream hessian2_codec repository (see
https://github.com/alibaba/hessian2-codec/pull/37). The patch, if
accepted will fix the problem for future versions of Envoy, but in 1.32
we are stuck with this issue.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>