Fix librdkafka Envoy dependency builds
New version of LLD coming with llvm-18 generates erros when building
librdkafka dependency on Envoy (see
https://github.com/confluentinc/librdkafka/issues/4593).

While there is an upstream PR fixing the issue, it has not been
merged and even when it will be merged, Envoy release v1.32 will still
be stuck with an old version without this fix.

I'm working around this with a patch to Envoy codebase by disabling
the use of linker scripts in librdkafka builds alltogether, and as
a consequence disabling the issues that come from the symbols defined
in the linker script only.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>