Replace calls to EC_group_p256 with EC_GROUP_new_by_curve_name
The problem is that in the version of BoringSSL that Envoy uses in
FIPS mode, EC_group_pX functions don't exist as they were introduced
later.

We can consider as an alternative just updating the BoringSSL version
to a newer one (especially given that the version we use is already
a few years old at this point), but that's a potentially larger
work item, so for now I'm just patching it.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>